+++ android/src/main/java/com/pspdfkit/react/events/CustomAnnotationContextualMenuItemTappedEvent.kt
@@ -1,3 +1,5 @@
+@file:Suppress("DEPRECATION", "OVERRIDE_DEPRECATION")
+
 package com.pspdfkit.react.events
 
 import androidx.annotation.IdRes