+++ android/src/main/java/com/pspdfkit/react/events/CustomTextSelectionContextualMenuItemTappedEvent.kt
@@ -1,3 +1,5 @@
+@file:Suppress("DEPRECATION", "OVERRIDE_DEPRECATION")
+
 package com.pspdfkit.react.events
 
 import androidx.annotation.IdRes
@@ -31,4 +33,3 @@ class CustomTextSelectionContextualMenuItemTappedEvent : Event<CustomTextSelecti
         var EVENT_NAME = "customTextSelectionContextualMenuItemTapped"
     }
 }
-